# openapi.yaml
openapi: 3.0.3
info:
  title: Checkwt
  version: 1.0.0
  description: Check your JWT and JWE

servers:
  - url: https://checkwt.io

paths:
  /decode:
    post:
      summary: Decode JWT
      tags:
        - Tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JWTRequest'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWTResponse'
        400:
          description: Bad request - token not readable
        404:
          description: Encryption key not found
  /keys/{keyId}:
    get:
      summary: Get signing key
      tags:
        - Tokens
      parameters:
        - name: keyId
          in: path
          required: true
          description: The ID of the signing key
          schema:
            type: string
            example: "019cfbfe-a43e-7c8d-b238-ffc538a933cc"
      responses:
        200:
          description: Key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyResponse'
        404:
          description: Signing key not found

components:
  schemas:
    JWTRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          example: eyJhbGciOiJFZERTQSIsImtpZCI6IjAxOWQwMTY0LTdjNGMtNzA4YS04MjU3LTA2ODhjYmFiODQ3NyIsInR5cCI6IkpXVCJ9.eyJjb3VudHJ5IjoiZ2xvYmFsIiwiZGVzY3JpcHRpb24iOiIiLCJlbnZpcm9ubWVudCI6InRlc3QiLCJleHAiOjE4MDUzODA4NTUsImlhdCI6MTc3Mzg0NDg1NSwicHJvZHVjdCI6Imdsb2JhbGN1c3RvbWVyYWNjb3VudCJ9.jajR9nID-jtfUU2-qZrMcXKW4KQFBBulXVnwSMFuGnZefE_BPtrHAPxhIilIghxM1GHDRxhXv11Oi6PfrC3ZCw

    JWTResponse:
      type: object
      required:
        - header
        - payload
        - signatureValid
      properties:
        header:
          type: object
        payload:
          type: object
        signatureValid:
          type: boolean
    KeyResponse:
      type: object
      required:
        - key
      properties:
        key:
          type: object
    SupportedAlgorithms:
      type: object
      description: Lists of supported algorithms for JWT (signing) and JWE (key management and content encryption).
      properties:
        jwtSigningAlgs:
          type: array
          items:
            type: string
          description: Supported `alg` values for JWT signing/verification.
          example: ["EdDSA", "ES256", "RS256", "PS256"]
        jweKeyManagementAlgs:
          type: array
          items:
            type: string
          description: Supported `alg` (key management) values for JWE.
          example: ["RSA-OAEP-256", "RSA-OAEP", "RSA1_5", "ECDH-ES", "ECDH-ES+A256KW", "A256KW"]
        jweContentEncryptionAlgs:
          type: array
          items:
            type: string
          description: Supported `enc` (content encryption) values for JWE.
          example: ["A256GCM", "A128GCM", "A128CBC-HS256"]